[General]: Update Copilot instructions (.github/copilot-instructions.md)
[Client]: Update client game and refresh integration (cl_clientgame.cpp, cl_refresh.cpp)

[CLGame]: Client-side game/system updates — core, parsing, packets, temp entities, and event/entity handling (clg_main.cpp, clg_client.cpp, clg_parse.cpp, clg_packet_entities.cpp, packet_entities/*, clg_events_entities.cpp, clg_temp_entities.cpp, clg_local.h)

[SharedGame]: Adjust player movement slide logic (sg_pmove_slidemove.cpp)

[SVGame]: Server-game updates — gamemode logic, weapons, edict pool, stepmove, and skeletal hitboxes
Game modes: svg_gm_cooperative.cpp, svg_gm_deathmatch.cpp, svg_gm_singleplayer.cpp
Weapons & edict pool: svg_player_weapon.cpp, svg_weapon.cpp, svg_edict_pool.cpp
Skeletal hitboxes & movement: svg_skeletal_hitboxes.cpp, svg_skeletal_hitboxes.h, svg_stepmove.cpp

+++ src/baseq2rtxp/svgame/svg_stepmove.cpp
@@ -314,9 +314,9 @@ M_ChangeYaw
 void M_ChangeYaw(svg_base_edict_t *ent)
 {
     // Get angle modded angles.
-    const float current = QM_AngleMod(ent->s.angles[YAW]);
+    const double current = QM_AngleMod(ent->s.angles[YAW]);
     // Get ideal desired for yaw angle.
-    const float ideal = ent->ideal_yaw;
+    const double ideal = ent->ideal_yaw;
 
     // If we're already facing ideal yaw, escape.
     if ( current == ideal ) {
